Refactor pools
